initial commit: small changes to progress bar
