--- src/Components/Organisms/ProgressBar/styles.module.scss
@@ -1,38 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;  
-    justify-content: center;
-    align-items: center;
-    height: 50vh; 
-  }
-  
-.title {
-    font-size: 32px;
-    font-weight: bold;
-    margin-bottom: 32px;
-    color: white;
-}
-  
-.progressContainer {
-    width: 1000px; 
-    background-color: #e0e0df;
-    border-radius: 5px;
-    overflow: hidden;
-}
-  
-.progressBar {
-    height: 24px;
-    border-radius: 5px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: white;
-    font-weight: bold;
-}
-
-.totalRaised {
-    font-size: 32px;
-    font-weight: bold;
-    margin-bottom: 32px;
-    color: white;
-}
